+++ src/com/github/dreamrec/IDataProvider.java
@@ -9,14 +9,14 @@ public interface IDataProvider {
     void stopRecording();
     double getIncomingDataFrequency();
     long getStartTime();
-    short ch1Poll();
+    int ch1Poll();
     int ch1Size();
-    short ch2Poll();
+    int ch2Poll();
     int ch2Size();
-    short acc1Poll();
+    int acc1Poll();
     int acc1Size();
-    short acc2Poll();
+    int acc2Poll();
     int acc2Size();
-    short acc3Poll();
+    int acc3Poll();
     int acc3Size();
 }
